% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funding.R
\name{add_funding}
\alias{add_funding}
\title{Add funding details on README and .Rmd files}
\usage{
add_funding(
  fund_message = "This work was funded by",
  fund_name,
  fund_code = NULL,
  html_text = NULL
)
}
\arguments{
\item{fund_message}{\code{character} message to be written}

\item{fund_name}{\code{character} name of the funding body}

\item{fund_code}{number of the grant/funding or \code{NULL} to ignore}

\item{html_text}{a \code{character} with HTML code to be inserted in the
funding area. If used, all other parameters are ignored.}
}
\value{
Edited files.
}
\description{
Add funding details on README and .Rmd files
}
\examples{
\dontrun{
add_funding(fund_name = "My funder", fund_code = "1000")
}
}
