% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gitrepo.R
\name{add_gitrepo}
\alias{add_gitrepo}
\title{Add GitHub repo address on README and .Rmd files}
\usage{
add_gitrepo(user_name, repo_name = NULL)
}
\arguments{
\item{user_name}{\code{character} name of GitHub account (usually your user name)}

\item{repo_name}{\code{character} name of the repository. Except in the cases your repository will be named differently then
the folder, you should leave it as \code{NULL}.}
}
\value{
Edited files.
}
\description{
Add GitHub repo address on README and .Rmd files
}
\examples{
\dontrun{
add_gitrepo("my-git-account")
}
}
