% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_institution.R
\name{add_institution}
\alias{add_institution}
\title{Add institutional details on README and .Rmd files}
\usage{
add_institution(
  inst_message = "Dataset edited by",
  inst_name,
  inst_details = NULL,
  html_text = NULL
)
}
\arguments{
\item{inst_message}{\code{character} message to be written}

\item{inst_name}{\code{character} name of the institution}

\item{inst_details}{\code{character} further details. Will be separated by a "|"}

\item{html_text}{a \code{character} with HTML code to be inserted in the
institutional details area. If used, all other parameters are ignored.}
}
\value{
Edited files.
}
\description{
Add institutional details on README and .Rmd files
}
\examples{
\dontrun{
add_institution(inst_name = "My institution")
}
}
