% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obisdi_create.R
\name{.obisdi_create}
\alias{.obisdi_create}
\title{Create an obisdi standard project}
\usage{
.obisdi_create(path, ...)
}
\arguments{
\item{path}{the path}

\item{...}{other parameters passed by the widget}
}
\value{
the directory
}
\description{
Create an obisdi standard project
}
\examples{
\dontrun{
obisdi_create("my_path", "My project")
}
}
\keyword{internal}
