% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_metadata.R
\name{gen_meta_keywords}
\alias{gen_meta_keywords}
\title{Generate metadata object for keywords}
\usage{
gen_meta_keywords(keyword, keywordThesaurus)
}
\arguments{
\item{keyword}{a \code{vector} containing one or various keywords}

\item{keywordThesaurus}{one single thesaurus for the keywords set}
}
\value{
a meta_keyword object containing the keywords section xml
}
\description{
Generate metadata object for keywords
}
\examples{
\dontrun{
gen_meta_keywords(c("Keyword 1", "Keyword 2"), "Thesaurus of botany")
}
}
