% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_metadata.R
\name{gen_meta_methods}
\alias{gen_meta_methods}
\title{Generate metadata object for methods info}
\usage{
gen_meta_methods(
  steps_description,
  studyExtent,
  samplingDescription,
  qualityControl = NULL
)
}
\arguments{
\item{steps_description}{a \code{vector} containing one or more step descriptions}

\item{studyExtent}{the extent of the study area}

\item{samplingDescription}{the description of the sampling}

\item{qualityControl}{the quality control procedure (optional)}
}
\value{
a meta_method object containing the methods section xml
}
\description{
Generate metadata object for methods info
}
\examples{
\dontrun{
gen_meta_methods("The description", "The study extent", "A sampling description")
}
}
