% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_metadata.R
\name{gen_meta_person}
\alias{gen_meta_person}
\title{Generate metadata object for individuals}
\usage{
gen_meta_person(
  givenName,
  surName,
  organizationName,
  positionName,
  role = NULL,
  phone = NULL,
  electronicMailAddress = NULL,
  deliveryPoint = NULL,
  postalCode = NULL,
  city = NULL,
  administrativeArea = NULL,
  country = NULL,
  onlineUrl = NULL
)
}
\arguments{
\item{givenName}{first name}

\item{surName}{family name}

\item{organizationName}{organization name}

\item{positionName}{position name}

\item{role}{role in the organization (optional)}

\item{phone}{phone (optional)}

\item{electronicMailAddress}{e-mail (optional, but recommended)}

\item{deliveryPoint}{address related (optional)}

\item{postalCode}{postal code (optional)}

\item{city}{city (optional)}

\item{administrativeArea}{administrative area (optional)}

\item{country}{country (ISO conforming, optional)}

\item{onlineUrl}{website or ORCID page of the person (optional)}
}
\value{
a list containing the original values and a list of XML ready objects
}
\description{
Generate metadata object for individuals
}
\details{
There is a + method for selecting just one or multiple names from a meta_person object (the object generated with this function).
Thus, suppose you create an object called \code{people} for three persons: John Doe, Max Planck, Carl Pol. But you decide that only one will be used to the \code{contact} field. Then you can do:

\code{selected_people <- people + "John Doe"}

Similarly you can supply a vector:

\code{selected_people <- people + c("John Doe", "Max Planck")}

Of course, names should match the original ones, otherwise an error is thrown.
}
\examples{
\dontrun{
people <- gen_meta_person("John", "Doe", "My company", "CEO")
people
}
}
