% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_metadata.R
\name{gen_meta_project}
\alias{gen_meta_project}
\title{Generate metadata object for project info}
\usage{
gen_meta_project(
  personnel = NULL,
  funding = NULL,
  studyAreaDescription = NULL,
  designDescription = NULL
)
}
\arguments{
\item{personnel}{an object of class \code{meta_person} generated with \link{gen_meta_person} optional}

\item{funding}{optional}

\item{studyAreaDescription}{optional}

\item{designDescription}{optional}
}
\value{
meta_proj object containing the project section xml
}
\description{
Generate metadata object for project info
}
\details{
You can chose any or several fields to fill.
}
\examples{
\dontrun{
gen_meta_project(funding = "Funder #0000")
}
}
