% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_metadata.R
\name{gen_meta_taxoncov}
\alias{gen_meta_taxoncov}
\title{Generate metadata object for taxonomic coverage}
\usage{
gen_meta_taxoncov(
  generalTaxonomicCoverage,
  taxonRankName = NULL,
  taxonRankValue = NULL,
  commonName = NULL
)
}
\arguments{
\item{generalTaxonomicCoverage}{a description of the taxonomic coverage}

\item{taxonRankName}{optional}

\item{taxonRankValue}{optional}

\item{commonName}{optional}
}
\value{
a meta_cov object containing the taxonomic coverage xml
}
\description{
Generate metadata object for taxonomic coverage
}
\examples{
\dontrun{
gen_meta_taxoncov("Plantae of Europe")
}
}
