% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_metadata.R
\name{gen_meta_temporalcov}
\alias{gen_meta_temporalcov}
\title{Generate metadata object for temporal coverage}
\usage{
gen_meta_temporalcov(singleDateTime = NULL, beginDate = NULL, endDate = NULL)
}
\arguments{
\item{singleDateTime}{a single date for the coverage}

\item{beginDate, endDate}{start (end) date of the coverage.}
}
\value{
a meta_cov object containing the temporal coverage xml
}
\description{
Generate metadata object for temporal coverage
}
\details{
You can supply a single date \code{singleDateTime} or the start AND end date,
but you can supply both the single date and the range.

In all cases use the format YYYY-MM-DD (ISO conform) or just YYYY-MM or YYYY
}
\examples{
\dontrun{
gen_meta_temporalcov("2023-05-05")
}
}
