% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_metadata.R
\name{gen_metadata}
\alias{gen_metadata}
\title{Generate an EML metadata file in the XML format to inclusion in the IPT}
\usage{
gen_metadata(
  outfolder = "data/processed",
  outfile = "metadata",
  title,
  creator,
  metadataProvider,
  language,
  metadataLanguage,
  abstract,
  associatedParty = NULL,
  coverage = NULL,
  purpose = NULL,
  contact = NULL,
  methods = NULL,
  project = NULL,
  intellectualRights = NULL,
  keywordSet = NULL
)
}
\arguments{
\item{outfolder}{the folder where the xml will be saved}

\item{outfile}{the name of the metadata file}

\item{title}{the title of the dataset}

\item{creator}{the creator of the dataset. Should be of meta_person class, created with \code{\link[=gen_meta_person]{gen_meta_person()}}}

\item{metadataProvider}{the provider of the dataset. Should be of meta_person class, created with \code{\link[=gen_meta_person]{gen_meta_person()}}}

\item{language}{the language of the dataset (ISO conform)}

\item{metadataLanguage}{the language of the metadata (ISO conform)}

\item{abstract}{abstract of the project (\code{character})}

\item{associatedParty}{associated parties. Should be of meta_person class, created with \code{\link[=gen_meta_person]{gen_meta_person()}}}

\item{coverage}{geographical, taxonomic or/and temporal coverage. Should be a list of objects of class meta_cov created using \code{\link[=gen_meta_geocov]{gen_meta_geocov()}}, \code{\link[=gen_meta_temporalcov]{gen_meta_temporalcov()}} or \code{\link[=gen_meta_taxoncov]{gen_meta_taxoncov()}}}

\item{purpose}{the purpose of the dataset}

\item{contact}{the contacts of the dataset. Should be of meta_person class, created with \code{\link[=gen_meta_person]{gen_meta_person()}}}

\item{methods}{methods of the project. Should be a meta_methods object created with \code{\link[=gen_meta_methods]{gen_meta_methods()}}.}

\item{project}{project info created with \code{\link[=gen_meta_project]{gen_meta_project()}}}

\item{intellectualRights}{intellectual rights (text)}

\item{keywordSet}{keywords. Should be a list of meta_keyword objects created with \code{\link[=gen_meta_keywords]{gen_meta_keywords()}}}
}
\value{
an XML file in the EML language
}
\description{
EML (Ecological Markup Language) is a standard format for metadata widely used by the biodiversity databases.
This function creates an standard EML file for use when submitting data through one IPT (\url{https://ipt.gbif.org/manual/en/ipt/latest/}).
}
\details{
Some of the fields need objects created with other functions of the "gen_meta" family. When the field asks for a list you can also supply a single object,
which will then be converted into a list.
For more information about the fields, consult the following link: \url{https://www.gbif.org/sites/default/files/gbif_resource/resource-80640/gbif_metadata_profile_guide_en_v1.pdf}

The & symbol should NOT be used in any text, otherwise the reading of the xml will fail.

When supplying names for the fields that require an object of type \code{meta_person}, you can select just a couple of names of the full object by using
the "+" or "-" method (for more details see \code{\link[=gen_meta_person]{gen_meta_person()}}).
}
\examples{
\dontrun{
person <- gen_meta_person("John", "Doe", "Researcher", "Institute")
gen_metadata(title = "My title", creator = person, metadataProvider = person,
language = "English", metadataLanguage = "English", abstract = "Some text.")
}
}
\seealso{
\code{\link[=gen_meta_person]{gen_meta_person()}},\code{\link[=gen_meta_geocov]{gen_meta_geocov()}}, \code{\link[=gen_meta_temporalcov]{gen_meta_temporalcov()}}, \code{\link[=gen_meta_taxoncov]{gen_meta_taxoncov()}}, \code{\link[=gen_meta_project]{gen_meta_project()}},
\code{\link[=gen_meta_methods]{gen_meta_methods()}}, \code{\link[=gen_meta_keywords]{gen_meta_keywords()}}
}
