% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obisdi-package.R
\docType{package}
\name{obisdi-package}
\alias{obisdi}
\alias{obisdi-package}
\title{obisdi: Ingest Marine Biodiversity Datasets for OBIS}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
