% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obisdi_create.R
\name{obisdi_create}
\alias{obisdi_create}
\title{Create an obisdi standard project}
\usage{
obisdi_create(path, dataset_name, type = "general")
}
\arguments{
\item{path}{the (absolute or relative) path to generate the project folder}

\item{dataset_name}{character with the dataset name}

\item{type}{type of project. At this moment, only "general" is available.}
}
\value{
An structured folder with the necessary files for data processing.
}
\description{
This function is used for...
}
\examples{
\dontrun{
obisdi_create("my_path", "My project")
}
}
