% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_metadata.R
\name{+.meta_person}
\alias{+.meta_person}
\title{Select name method for the meta_person class}
\usage{
\method{+}{meta_person}(e1, e2)
}
\arguments{
\item{e1}{object}

\item{e2}{names}
}
\value{
a new object meta_person
}
\description{
Select name method for the meta_person class
}
