% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_metadata.R
\name{print.meta_cov}
\alias{print.meta_cov}
\title{Print method for the generate_*cov group}
\usage{
\method{print}{meta_cov}(x)
}
\arguments{
\item{x}{}
}
\value{
printed object
}
\description{
Print method for the generate_*cov group
}
