% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_metadata.R
\name{print.meta_person}
\alias{print.meta_person}
\title{Print method for the generate_meta_person}
\usage{
\method{print}{meta_person}(x)
}
\arguments{
\item{x}{object}
}
\value{
printed object
}
\description{
Print method for the generate_meta_person
}
