% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obisdi-package.R
\docType{package}
\name{obisdi-package}
\alias{obisdi}
\alias{obisdi-package}
\title{obisdi: Ingest Marine Biodiversity Datasets for OBIS}
\description{
Create a structure for projects of data ingestion into OBIS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://iobis.github.io/obisdi}
  \item \url{https://github.com/iobis/obisdi}
  \item Report bugs at \url{https://github.com/iobis/obisdi/issues}
}

}
\author{
\strong{Maintainer}: Silas C. Principe \email{s.principe@unesco.org} (\href{https://orcid.org/0000-0002-8059-6304}{ORCID})

}
\keyword{internal}
